# Speed up lodestar unit test loading