--- a/packages/lodestar/test/unit/chain/factory/attestation/data.test.ts
+++ b/packages/lodestar/test/unit/chain/factory/attestation/data.test.ts
@@ -1,5 +1,5 @@
 import {expect} from "chai";
-import {config} from "@chainsafe/lodestar-config/mainnet";
+import {config} from "@chainsafe/lodestar-config/minimal";
 import {assembleAttestationData} from "../../../../../src/chain/factory/attestation/data";
 import {generateState} from "../../../../utils/state";
 import {generateEmptyBlock} from "../../../../utils/block";