--- a/packages/lodestar/tsconfig.test.json
+++ b/packages/lodestar/tsconfig.test.json
@@ -7,6 +7,9 @@
       "./node_modules/@types",
       "./types"
     ],
+    "declaration": false,
+    "declarationMap": false,     
+    "pretty": false,
     "strict": false
   }
 }