--- a/packages/lodestar/tsconfig.test.json
+++ b/packages/lodestar/tsconfig.test.json
@@ -11,5 +11,9 @@
     "declarationMap": false,     
     "pretty": false,
     "strict": false
+  },
+  "ts-node": {
+    "transpileOnly": true,
+    "files": true
   }
 }