Before:
```
real    1m42,309s
user    2m3,016s
sys     0m1,695s
```

After:
```
real    0m47,498s
user    0m59,736s
sys     0m1,254s
```